Features

+++ screensharing-service/Entities/ScreenMirroring/MousePositionChangeEvent.cs
@@ -2,7 +2,7 @@ namespace screensharing_service.Entities.ScreenMirroring
 {
     public class MousePosition:ScreenMirroringEvent
     {
-        public int top { get; set; }
-        public int left { get; set; }
+        public float top { get; set; }
+        public float left { get; set; }
     }
 }
